requirements.txt
cryptography
pyjwt[crypto]